+++ middleware/example_test.go
@@ -35,7 +35,7 @@ func Example_getCspNonce() {
 	l := log.New(os.Stdout, 100)(context.Background())
 	handler := middleware.Get(
 		loginHandler(),
-		middleware.WithOpts("example.com", 443, "secretKey", middleware.DirectIpStrategy, l),
+		middleware.WithOpts("example.com", 443, "super-h@rd-Pa$1word", middleware.DirectIpStrategy, l),
 	)
 	_ = handler // use handler
 
@@ -46,7 +46,7 @@ func Example_getCsrfToken() {
 	l := log.New(os.Stdout, 100)(context.Background())
 	handler := middleware.Get(
 		welcomeHandler(),
-		middleware.WithOpts("example.com", 443, "secretKey", middleware.DirectIpStrategy, l),
+		middleware.WithOpts("example.com", 443, "super-h@rd-Pa$1word", middleware.DirectIpStrategy, l),
 	)
 	_ = handler // use handler
 
@@ -55,7 +55,7 @@ func Example_getCsrfToken() {
 
 func ExampleGet() {
 	l := log.New(os.Stdout, 100)(context.Background())
-	opts := middleware.WithOpts("example.com", 443, "secretKey", middleware.DirectIpStrategy, l)
+	opts := middleware.WithOpts("example.com", 443, "super-h@rd-Pa$1word", middleware.DirectIpStrategy, l)
 	handler := middleware.Get(loginHandler(), opts)
 	_ = handler // use handler
 
@@ -64,7 +64,7 @@ func ExampleGet() {
 
 func ExampleAll() {
 	l := log.New(os.Stdout, 100)(context.Background())
-	opts := middleware.WithOpts("example.com", 443, "secretKey", middleware.DirectIpStrategy, l)
+	opts := middleware.WithOpts("example.com", 443, "super-h@rd-Pa$1word", middleware.DirectIpStrategy, l)
 
 	myHandler := http.HandlerFunc(
 		func(w http.ResponseWriter, _ *http.Request) {
@@ -85,7 +85,7 @@ func ExampleWithOpts() {
 	opts := middleware.WithOpts(
 		"example.com",
 		443,
-		"secretKey",
+		"super-h@rd-Pa$1word",
 		// assuming your application is deployed behind cloudflare.
 		middleware.SingleIpStrategy("CF-Connecting-IP"),
 		l,