+++ server/server_test.go
@@ -24,7 +24,7 @@ import (
 )
 
 func getSecretKey() string {
-	key := "hard-password"
+	key := "super-h@rd-Pa$1word"
 	return key
 }